refactor(services/contest): split change dataclasses into changes module
state.py mixed two concerns: the change-shape data classes
(ChangeType, FieldChange, ResourceChange, ContestChangeSet) and the
ContestStateComparator service that produces them. Callers that only
needed to consume change shapes were dragging in the comparator and
its API-client dependency.

Move the four data classes into a new dom/core/services/contest/changes.py
and update the five importers (two in core, three in tests) to import
the types directly from changes.py and the comparator from state.py.

state.py: 382 → 291 LOC. No behavior change; ContestStateComparator
internals untouched.
