+++ dom/core/services/contest/state.py
@@ -1,9 +1,13 @@
 """Contest state comparison and change detection service."""
 
-from dataclasses import dataclass
-from enum import Enum
 from typing import Any
 
+from dom.core.services.contest.changes import (
+    ChangeType,
+    ContestChangeSet,
+    FieldChange,
+    ResourceChange,
+)
 from dom.core.services.protocols import DomJudgeAPIProtocol
 from dom.logging_config import get_logger
 from dom.types.config.processed import ContestConfig
@@ -11,89 +15,6 @@ from dom.utils.cli import get_secrets_manager
 from dom.utils.hashing import generate_team_username
 
 logger = get_logger(__name__)
-
-
-class ChangeType(str, Enum):
-    """Types of changes that can be detected."""
-
-    CREATE = "create"
-    UPDATE = "update"
-    NO_CHANGE = "no_change"
-
-
-@dataclass
-class FieldChange:
-    """Represents a change in a specific field."""
-
-    field: str
-    old_value: Any
-    new_value: Any
-
-    def __str__(self) -> str:
-        """Format change for display."""
-        return f"{self.field}: {self.old_value} → {self.new_value}"
-
-
-@dataclass
-class ResourceChange:
-    """Represents changes in contest resources (problems/teams)."""
-
-    resource_type: str  # "problems" or "teams"
-    to_add: list[str]  # IDs/names to add
-    to_remove: list[str]  # IDs/names to remove (not implemented yet)
-    unchanged: list[str]  # IDs/names that exist
-
-    @property
-    def has_changes(self) -> bool:
-        """Check if there are any changes."""
-        return bool(self.to_add or self.to_remove)
-
-    def __str__(self) -> str:
-        """Format resource changes for display."""
-        parts = []
-        if self.to_add:
-            parts.append(f"+{len(self.to_add)} to add")
-        if self.to_remove:
-            parts.append(f"-{len(self.to_remove)} to remove")
-        if self.unchanged:
-            parts.append(f"={len(self.unchanged)} unchanged")
-        return f"{self.resource_type}: {', '.join(parts) if parts else 'no changes'}"
-
-
-@dataclass
-class ContestChangeSet:
-    """Represents all detected changes for a contest."""
-
-    contest_shortname: str
-    change_type: ChangeType
-    field_changes: list[FieldChange]
-    resource_changes: list[ResourceChange]
-
-    @property
-    def has_changes(self) -> bool:
-        """Check if there are any changes at all."""
-        return (
-            self.change_type != ChangeType.NO_CHANGE
-            or bool(self.field_changes)
-            or any(rc.has_changes for rc in self.resource_changes)
-        )
-
-    def summary(self) -> str:
-        """Get a human-readable summary of changes."""
-        if self.change_type == ChangeType.CREATE:
-            return f"[green]CREATE[/green] contest '{self.contest_shortname}'"
-
-        if not self.has_changes:
-            return f"[dim]NO CHANGES[/dim] for contest '{self.contest_shortname}'"
-
-        parts = []
-        if self.field_changes:
-            parts.append(f"{len(self.field_changes)} field(s)")
-        for rc in self.resource_changes:
-            if rc.has_changes:
-                parts.append(rc.resource_type)
-
-        return f"[yellow]UPDATE[/yellow] contest '{self.contest_shortname}': {', '.join(parts)}"
 
 
 class ContestStateComparator:
